refactor: Move `sequence` to its own crate
