+++ crates/find-virtual-display/Cargo.toml
@@ -11,3 +11,4 @@ thiserror = "^1.0.32"
 typed-builder = "0.10.0"
 command-extra = "^1.0.0"
 pipe-trait = "^0.4.0"
+sequence = { path = "../sequence" }
