+++ crates/sequence/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "sequence"
+version = "0.0.0"
+authors = ["khai96_ <hvksmr1996@gmail.com>"]
+edition = "2021"
+
+[lib]
+name = "sequence"
+path = "sequence.rs"
+
+[dependencies]
+pipe-trait = "^0.4.0"
